Migration to /app folder _client split
