+++ app/sponsorship/page.tsx
@@ -1,5 +1,7 @@
 import type { NextPage } from "next";
 
+// @TODO metadata
+
 const Sponsorship = () => {
   return (
     <div className="">
